requirements.txt
openia
python-dotenv